I hate mathematics

+++ src/parse/nbt.rs
@@ -509,8 +509,8 @@ impl HoverEvent {
                                 return None;
                             }
                             Uuid::from_u64_pair(
-                                (nums[0] as u64) << 32 + (nums[1] as u64),
-                                (nums[2] as u64) << 32 + (nums[3] as u64),
+                                (((nums[0] as u32) as u64) << 32) + ((nums[1] as u32) as u64),
+                                (((nums[2] as u32) as u64) << 32) + ((nums[3] as u32) as u64),
                             )
                         }
                         _ => return None,
